refactor: break circular import between streaming and errorHandling
src/api/streaming.ts imports helpers from src/contexts/messages/
errorHandling.ts, which in turn imported STREAM_MESSAGE_TYPES from
streaming.ts — a direct cycle. Module resolution happened to hold up
but the dependency direction was inverted (api layer depending on
contexts layer).

Extracts the constants to src/api/streaming-types.ts. streaming.ts
re-exports them so existing consumers (IncomingMsg, MessagesContext,
responseParser) don't change. errorHandling.ts now depends only on
the pure-constants module.
