+++ src/api/streaming-types.ts
@@ -1,0 +1,9 @@
+export const STREAM_MESSAGE_TYPES = {
+  CONVERSATION_START: "conversation_start",
+  FINAL_RESPONSE: "final_response",
+  RUN_START: "run_start",
+  RUNNING: "running",
+  COMPLETED: "completed",
+  MESSAGE_PART: "message_part",
+  ERROR: "error",
+} as const;
